+++ src/main/scala/org/podval/tools/test/framework/JUnit4ScalaNative.scala
@@ -1,7 +1,6 @@
 package org.podval.tools.test.framework
 
 import org.podval.tools.build.scalanative.ScalaNativeBackend
-import org.podval.tools.build.ScalaDependency
 
 object JUnit4ScalaNative extends FrameworkDescriptor(
   name = "Scala Native JUnit test framework",
@@ -11,9 +10,9 @@ object JUnit4ScalaNative extends FrameworkDescriptor(
   versionDefault = ScalaNativeBackend.versionDefault,
   className = "scala.scalanative.junit.JUnitFramework",
   sharedPackages = List("com.novocode.junit", "junit.framework", "junit.extensions", "org.junit"),
-  usesTestSelectorAsNestedTestSelector = true,
+  usesTestSelectorAsNestedTestSelector = true
+):
   // This is a Scala Native-only test framework
-  forJVM = ForBackend.notSupported,
-  forJS = ForBackend.notSupported
-) with ScalaDependency.Maker
-
+  override val forJVM   : None.type = None
+  override val forJS    : None.type = None
+