+++ src/main/scala/org/podval/tools/test/framework/JUnit5.scala
@@ -1,6 +1,6 @@
 package org.podval.tools.test.framework
 
-import org.podval.tools.build.{JavaDependency, Version}
+import org.podval.tools.build.Version
 
 // https://github.com/sbt/sbt-jupiter-interface
 // https://github.com/sbt/sbt-jupiter-interface/blob/master/src/library/src/main/java/net/aichler/jupiter/api/JupiterFramework.java
@@ -14,9 +14,9 @@ object JUnit5 extends FrameworkDescriptor(
   sharedPackages = List("com.github.sbt.junit.jupiter.api", "org.junit"),
   tagOptionStyle = OptionStyle.ListWithEq,
   includeTagsOption = "--include-tags",
-  excludeTagsOption = "--exclude-tags",
+  excludeTagsOption = "--exclude-tags"
+):
   // JUnit5 is not supported since it uses its own test detection mechanism.
-  forJVM = ForBackend.notSupported,
-  forJS = ForBackend.notSupported,
-  forNative = ForBackend.notSupported
-) with JavaDependency.Maker
+  override val forJVM   : None.type = None
+  override val forJS    : None.type = None
+  override val forNative: None.type = None